+++ database/migrations/2019_06_6_145506_create_article_tag_table.php
@@ -14,12 +14,12 @@ class CreateArticleTagTable extends Migration
     public function up()
     {
         Schema::create('article_tag', function (Blueprint $table) {
-            $table->bigIncrements('id');
+            $table->increments('id');
             $table->timestamp('created_at');
             $table->timestamp('updated_at');
-            $table->unsignedBigInteger('article_id')->nullable();
+            $table->unsignedInteger('article_id')->nullable();
             $table->foreign('article_id')->references('id')->on('articles');
-            $table->unsignedBigInteger('tag_id')->nullable();
+            $table->unsignedInteger('tag_id')->nullable();
             $table->foreign('tag_id')->references('id')->on('tags');
         });
     }
